Match `func_80055140_55D40`, `func_800557A0_563A0`
func_80055140: set sprite texture/anim fields (unk_88 masked, unk_0C, unk_8A,
unk_8C) then sync via func_80052E14. Corrected its prototype's group arg to s16.

func_800557A0: kill every loaded texture descriptor in D_800C9530_CA130 via
HuSprKill (mirrors func_80057158).

+++ include/game/sprite.h
@@ -115,7 +115,7 @@ void func_80054FF8_55BF8(s16 group, s16 member, s32 arg2);
 void func_80055024_55C24(s16 group, s16 member, s16 arg2, s32 arg3);
 void func_800550B4_55CB4(s16 group, s16 member, f32 arg2);
 void func_800550F4_55CF4(s16, s16, s32);
-void func_80055140_55D40(s32, s16, u16, s32);
+void func_80055140_55D40(s16, s16, u16, s32);
 u8 func_800551AC_55DAC(s16 group, s16 member);
 HuSprite_Unk84_Struct* func_80055194_55D94(s16 arg0);
 void HuSprScaleSet(s16 group, s16 member, f32 x, f32 y);
